src/requirements.txt
langgraph>=0.0.10
langchain-core>=0.1.0
langchain-openai>=0.0.2
python-dotenv>=1.0.0 